refactor: fix import cycle
