+++ cyberdrop_dl/cli/__init__.py
@@ -1,6 +1,5 @@
 import datetime
 from collections.abc import Iterable
-from enum import auto
 from pathlib import Path
 from typing import Annotated, Any, Literal, Self
 
@@ -9,19 +8,8 @@ from cyclopts.core import App
 from pydantic import BaseModel, Field, computed_field, field_validator, model_validator
 
 from cyberdrop_dl import __version__
-from cyberdrop_dl.compat import CIStrEnum
 from cyberdrop_dl.models.types import HttpURL
-
-
-class UIOptions(CIStrEnum):
-    DISABLED = auto()
-    ACTIVITY = auto()
-    SIMPLE = auto()
-    FULLSCREEN = auto()
-
-    @property
-    def is_disabled(self) -> bool:
-        return self is UIOptions.DISABLED
+from cyberdrop_dl.progress import UIOptions
 
 
 @Parameter(name="*")
